==> notebooks/1 - finding genres.ipynb <==
import requests
from bs4 import BeautifulSoup
from time import sleep
import json

base_url = 'https://api.mobygames.com/v1/'
endpoint_var = 'genres'
api_key = open("../data/mobygames_api_key.txt")
endpoint= base_url + endpoint_var

parameters = {"format":"normal", 
              "api_key": api_key 
#               "genre":1, 
#               "id":2
             }

### Endpoints

We will be using the `requests` package to make a `GET` request to an API. Similar to webscraping, APIs require an endpoint to tell python where to send the request.


response = requests.get(endpoint, params=parameters)
response

## jsonify the response

res = response.json()

res

#this is a list
type(res["genres"])

#each list item is a dictionary
type(res["genres"][0])

res["genres"][3]

#initialize easier genre dict
genre_Name = []
genre_ID = []

## iterate through list and pull out genre names and genre ids
for dictionary in res["genres"]:
    genre_Name.append(dictionary["genre_name"])
    genre_ID.append(dictionary["genre_id"])
    
    

#list comprehension
genre_Name = [name.lower() for name in genre_Name]

genre_Name

## zipping lists together to create easier genre dict
genre_dictionary = dict(zip(genre_Name, genre_ID))

genre_dictionary.keys()

for (key, value) in genre_dictionary.items():
    if "simulation" in key:
        print(value)

### use params

base_url = 'https://api.mobygames.com/v1/'
endpoint_var = 'games'
api_key = open("../data/mobygames_api_key.txt")
endpoint= base_url + endpoint_var

parameters = {"format":"normal", "api_key": api_key, "genre":1, "offset":100}

response = requests.get(endpoint, params=parameters)

response

res2 = response.json()

len(res2["games"])

res2["games"][0].keys()

with open('../data/json_test.json', 'w') as f:
    json.dump(res2, f)
