==> notebooks/3 - cleaning_vehicle_sim_genre_df.ipynb <==
import pandas as pd

df = pd.read_csv("../data/vehicle_sim_158_df_w_genre.csv")

df

### checking for duplicates (really should not be any)

#from https://stackoverflow.com/questions/29836836/how-do-i-filter-a-pandas-dataframe-based-on-value-counts
df_dupes = df[df['game_id'].map(df['game_id'].value_counts()) > 1]

#this shows no duplicates by game_id
df_dupes

### creating a "all genres" column

df["title"].astype(str)

df.info()

df["all_genres"] = df["genre_name1"].map(str) + " , "+	df["genre_name2"].map(str) + " , "+	df["genre_name3"].map(str) + " , "+	df["genre_name4"].map(str) + " , "+	df["genre_name5"].map(str) + " , "+	df["genre_name6"].map(str) + " , "+	df["genre_name7"].map(str) + " , "+	df["genre_name8"].map(str) + " , "+	df["genre_name9"].map(str) + " , "+	df["genre_name10"].map(str) + " , "+	df["genre_name11"].map(str) + " , "+	df["genre_name12"].map(str) + " , "+	df["genre_name13"].map(str) + " , "+	df["genre_name14"].map(str) + " , "+	df["genre_name15"].map(str) + " , "+	df["genre_name16"].map(str) + " , "+	df["genre_name17"].map(str) + " , "+	df["genre_name18"].map(str) + " , "+	df["genre_name19"].map(str) + " , "+	df["genre_name20"].map(str) + " , "+	df["genre_name21"].map(str) + " , "+	df["genre_name22"].map(str) + " , "+	df["genre_name23"].map(str) + " , "+	df["genre_name24"].map(str) + " , "+	df["genre_name25"].map(str)


df["all_genres"]

df

df.to_csv("../data/genre_158_with_allgenre_column_11Dec21.csv")

# note: I'm not going to do much genre filtering - I'm just going with whatever is tagged, at least once, as "vehicle simulator"

### finding earliest release date for each game

df1 = df.loc[:, df.columns.str.contains('release_date')].astype('datetime64[s]')

df1

#after discussing with teachers, going to use release_date1 only, instead of comparing - also it seems like release_date1 is actually the earliest one, almost every time, or its imprecise and there would be no real way to compare

### changing release_date1 to datetime and creating new column of just the year

df["release_date1"] = df["release_date1"].astype('datetime64[s]')

df["release_date1_year"] = df["release_date1"].dt.year

df

df.to_csv("../data/genre_158_release_year_11Dec21_2.csv")
