==> notebooks/4 - getting_company_info.ipynb <==
import requests
from bs4 import BeautifulSoup
from time import sleep
import pandas as pd
import json
import pickle

df = pd.read_csv("../data/genre_158_only1releaseYear_NoGenre_11Dec21.csv")

df

### testing and experimentation

url="https://api.mobygames.com/v1/games/276/platforms/3"
api_key = open("../data/mobygames_api_key.txt")
parameters = {"format":"normal", "api_key": api_key}

response = requests.get(url, params=parameters)
response

res = response.json()

len(res["releases"])

#shows all release dates for a response
for item in res["releases"]:
    for key, value in item.items():
        if key == "release_date":
            print(value)

# note: res["releases"] is a list. The list contains dictionaries. 
# Each dictionary contains info about a release for a particular date, which companies were involved in the release
# and which countries it was released in
# this is actually a very rich source of info
# The list can be of variable length, depending on how many releases there were
# I am choosing the first item in this list always, as it seems to often be the first release, period.
# however, there is a lot one could do with all of the different releases - could look to see which countries appear
# most often across all releases, regardless of time
# or you could see if there is correlation between certain countries and certain time periods
# I lack the time for such analysis, but wanted to make a note

### creating the loop for scraping

# DO HERE: TURN COLUMNS BACK INTO LISTS
# ZIP THEM TOGETHER - https://www.delftstack.com/howto/python/how-to-loop-through-multiple-lists-in-python/
#pandas stuff - to replace the 2 below
game_id_list = df["game_id"].tolist()
platform_id_list = df["platform_id1"].tolist()
# game_id_list = [5, 107]
# platform_id_list = [2,2]


url_list = []
results_list = []

# lists containing game ids and platform ids for the loop, and list of final urls to submit
for a,b in zip(game_id_list, platform_id_list):
    url_list.append("https://api.mobygames.com/v1/games/" + str(a) + "/platforms/" + str(b))

print(len(url_list))
# print(url_list)

results_list = []

# NEXT STEP - PUMP THIS LIST ABOVE THROUGH A LOOP FOR SCRAPING
for url in unfinished_url_list:
    api_key = open("../data/mobygames_api_key.txt")
    parameters = {"format":"normal", 
                  "api_key": api_key}
    response = requests.get(url, params=parameters)
    print(response)
    res = response.json()
    results_list.append(res)
    sleep(10.03)

results_list[-1]

# WHEN YOU RETURN, RUN CELL BELOW THIS!!


f = open("../data/platform_1119_to_end.pkl","wb")
pickle.dump(results_list,f)
f.close()

#the final one that got through
results_list[-1]

url_list[874]

unfinished_url_list = unfinished_url_list[244:]

unfinished_url_list[0:5]

### Next step: actually extra info you want from the local files

# starting point - you will have to adjust
# LOAD LOCAL FILES, ADD THEM TOGETHER INTO ONE LIST


#shows all release dates for a response
for item in res["releases"]:
    for key, value in item.items():
        if key == "release_date":
            print(value)
