==> notebooks/5 - getting companies from local files.ipynb <==
import pandas as pd
import pickle
import numpy as np

with open('../data/platform_to_875.pkl', 'rb') as f1:
    scrape1 = pickle.load(f1)
with open('../data/platform_875_to_1119.pkl', 'rb') as f2:
    scrape2 = pickle.load(f2)
with open('../data/platform_1119_to_end.pkl', 'rb') as f3:
    scrape3 = pickle.load(f3)

type(scrape1)

len(scrape2)

len(scrape3)

data = []

data.extend(scrape1)

type(data[0])

data.extend(scrape2)

type(data[1])

data.extend(scrape3)

len(data)

f = open("../data/full_company_scrape.pkl","wb")
pickle.dump(data,f)
f.close()

### Ok, now we work with the full file (use full_company_scrape.pkl)

with open('../data/full_company_scrape.pkl', 'rb') as f:
    data = pickle.load(f)

# things to get
data[0]

#this can be multiple lengths, is a list
data[0]["releases"]

# for companies, list of dicts, each dict is a different company with a different role - need to check through this
# this is probably variable in length, depending on how many companies are credited in different roles
# data[0]["releases"][0]["companies"]

# for one company with one role - company id, company name, and role
# data[0]["releases"][0]["companies"][0]

# this might be a logical condition to check, if the following equals "Published by", then put in published list
# data[0]["releases"][0]["companies"][0]["role"]

all_cos_for_game = []

# goal: print out all the different kinds of roles
for dictionary in data:
    all_cos_for_game.append(dictionary["releases"][0]["companies"])
#     for company in first_release_companies_list:
#         print(type(company))
#         print(company.keys())
#         list_of_dicts.append(company)
#         for key, value in company.items():
#             print(value)
#             if company["role"] == "Published by":
#                 publisher_name.append(company["company_name"])
#                 publisher_id.append(company["company_id"])
# #             else:
# #                 publisher_name.append("")
# #                 publisher_id.append("")
#             print(type(company.values()))
#             if company[key] == "Published by":
#                 publisher_name.append(company["company_name"])
#                 publisher_id.append(company["company_id"])
#             else:
#                 publisher_name.append("")
#                 publisher_id.append("")


len(all_cos_for_game)

# df = pd.DataFrame()
pub_results = []
# publishers = []
# developers = []
# pub_and_dev={"publisher_name":np.nan, "publisher_id":np.nan, "developer_name":np.nan, "developer_id":np.nan}

# def find_pub(dictionary):
#     if "Published by" in dictionary.values():
#         return dictionary
# def find_dev(dictionary):
#     if "Developed by" in dictionary.values():
#         return dictionary

# def get_data_from_dict(input_list):
#     # check every dictionary in the list to find the one that has pub and the one that has dev
#     for dictionary in input_list:
#         if dictionary["role"] == "Published by":
#             publishers.append(dictionary)
#     for dictionary in input_list:
#         if dictionary["role"] == "Developed by":
#             developers.append(dictionary)
        
#         publisher_dict = find_pub(dictionary)
#         developer_dict = find_dev(dictionary)
#     return publisher_dict, developer_dict
    
#     if (publisher_dict != None) and (developer_dict != None): #both
#         output_dict["publisher_name"] = publisher_dict["company_name"]
#         output_dict["publisher_id"] = publisher_dict["company_id"]
#         output_dict["developer_name"] = developer_dict["company_name"]
#         output_dict["developer_id"] = developer_dict["company_id"]
#         return output_dict
#     elif (publisher_dict == None) and (developer_dict != None): #dev only
#         output_dict["publisher_name"] = np.nan
#         output_dict["publisher_id"] = np.nan
#         output_dict["developer_name"] = developer_dict["company_name"]
#         output_dict["developer_id"] = developer_dict["company_id"]
#         return output_dict
#     elif (publisher_dict != None) and (developer_dict == None): # pub only
#         output_dict["publisher_name"] = publisher_dict["company_name"]
#         output_dict["publisher_id"] = publisher_dict["company_id"]
#         output_dict["developer_name"] = np.nan
#         output_dict["developer_id"] = np.nan
#         return output_dict
#     elif (publisher_dict == None) and (developer_dict == None): #neither
#         output_dict["publisher_name"] = np.nan
#         output_dict["publisher_id"] = np.nan
#         output_dict["developer_name"] = np.nan
#         output_dict["developer_id"] = np.nan
#         return output_dict


#OMG I finally got it - this here works - 14 Dec 21
for list_of_cos_for_each_game in all_cos_for_game:
    values_list = []
    publishers = []
    for dictionary in list_of_cos_for_each_game:
        for key, value in dictionary.items():
            values_list.append(value)
    if "Published by" in values_list:
        for dictionary in list_of_cos_for_each_game:
            if dictionary["role"] == "Published by":
                publishers.append(dictionary)
                values_list=[]
    pub_results.append(publishers)
#     elif "Published by" not in values_list:
#         print("elif")
#         publishers.append("")
        
#         if dictionary["role"] == "Published by":
#             publishers.append(dictionary)
#     for dictionary in list_of_cos_for_each_game:
#         if dictionary["role"] == "Developed by":
#             developers.append(dictionary)

#     if pubcheck == True:
# #         haspub +=1
#         for dictionary in item:
#             for key, value in dictionary.items():
#                 if value == "Published by":
#                     publisher_name.append(dictionary["company_name"])
#                     publisher_id.append(dictionary["company_id"])
#     elif pubcheck == False:
#         publisher_name.append("")
#         publisher_id.append("")
# #         nopub +=1

len(pub_results)

#this level contains a list of 0, 1, 2, or more publishers
pub_results[4]

### this stuff just to see breakdown of no pub, 1 pub, and more than 1 pub

no_pub = []
one_pub = []
more_than_1_pub = []


for item in pub_results:
    if len(item) == 0:
        no_pub.append(item)
    elif len(item) == 1:
        one_pub.append(item)
    elif len(item) > 1:
        more_than_1_pub.append(item)

print(len(no_pub))
print(len(one_pub))
print(len(more_than_1_pub))

### getting data into dataframe

#lists for columns
publisher1_name = []
publisher1_id = []
publisher2_name = []
publisher2_id = []

for item1 in pub_results:
    if len(item1) == 0:
        publisher1_name.append("")
        publisher1_id.append("")
        publisher2_name.append("")
        publisher2_id.append("")
    elif len(item1) == 1:
        publisher1_name.append(item1[0]["company_name"])
        publisher1_id.append(item1[0]["company_id"])
        publisher2_name.append("")
        publisher2_id.append("")
    elif len(item1)== 2:
        publisher1_name.append(item1[0]["company_name"])
        publisher1_id.append(item1[0]["company_id"])
        publisher2_name.append(item1[1]["company_name"])
        publisher2_id.append(item1[1]["company_id"])

print(len(publisher1_name))
print(len(publisher1_id))
print(len(publisher2_name))
print(len(publisher2_id))

#section for getting column lists into df
main_list = []
main_list.append(publisher1_name)
main_list.append(publisher1_id)
main_list.append(publisher2_name)
main_list.append(publisher2_id)

publisher_df = pd.DataFrame(main_list).transpose()
publisher_df.columns = ["publisher1_name", "publisher1_id", "publisher2_name", "publisher2_id"]

publisher_df.to_csv("../data/publisher_data_14Dec21.csv")

### doing it all again for developer

dev_result = []

for list_of_cos_for_each_game in all_cos_for_game:
    values_list = []
    developers = []
    for dictionary in list_of_cos_for_each_game:
        for key, value in dictionary.items():
            values_list.append(value)
    if "Developed by" in values_list:
        for dictionary in list_of_cos_for_each_game:
            if dictionary["role"] == "Developed by":
                developers.append(dictionary)
                values_list=[]
    dev_result.append(developers)

len(dev_result)

### checking to see max num of devs

no_pub = []
one_pub = []
more_than_1_pub = []

for item in dev_result:
    if len(item) == 0:
        no_pub.append(item)
    elif len(item) == 1:
        one_pub.append(item)
    elif len(item) > 2:
        more_than_1_pub.append(item)

print(len(no_pub))
print(len(one_pub))
print(len(more_than_1_pub))

print(len(more_than_1_pub[0]))

#result: max lenght is 3, but it is only one game, and I do not care (I think it is a sonic game)

### getting data into dataframe

#lists for columns
publisher1_name = []
publisher1_id = []
publisher2_name = []
publisher2_id = []

for item1 in dev_result:
    if len(item1) == 0:
        publisher1_name.append("")
        publisher1_id.append("")
        publisher2_name.append("")
        publisher2_id.append("")
    elif len(item1) == 1:
        publisher1_name.append(item1[0]["company_name"])
        publisher1_id.append(item1[0]["company_id"])
        publisher2_name.append("")
        publisher2_id.append("")
    elif len(item1) == 2:
        publisher1_name.append(item1[0]["company_name"])
        publisher1_id.append(item1[0]["company_id"])
        publisher2_name.append(item1[1]["company_name"])
        publisher2_id.append(item1[1]["company_id"])
    elif len(item1) == 3:
        publisher1_name.append(item1[0]["company_name"])
        publisher1_id.append(item1[0]["company_id"])
        publisher2_name.append(item1[1]["company_name"])
        publisher2_id.append(item1[1]["company_id"])

print(len(publisher1_name))
print(len(publisher1_id))
print(len(publisher2_name))
print(len(publisher2_id))

main_list = []
main_list.append(publisher1_name)
main_list.append(publisher1_id)
main_list.append(publisher2_name)
main_list.append(publisher2_id)

dev_df = pd.DataFrame(main_list).transpose()


dev_df.columns = ["dev1_name", "dev1_id", "dev2_name", "dev2_id"]

dev_df.to_csv("../data/dev_data_14Dec21.csv")
