==> notebooks/6 - trying to simply scrape website instead.ipynb <==
import requests
from bs4 import BeautifulSoup  
import pandas as pd
import re
import soupsieve as sv
from time import sleep

moby_urls = ["http://www.mobygames.com/game/indianapolis-500-the-simulation", 
"http://www.mobygames.com/game/test-drive",
"http://www.mobygames.com/game/moto-racer", 
"http://www.mobygames.com/game/test-drive-iii-the-passion", 
"http://www.mobygames.com/game/star-wars-episode-i-racer", 
"http://www.mobygames.com/game/midtown-madness", 
"http://www.mobygames.com/game/driver", 
"http://www.mobygames.com/game/test-drive-5", 
"http://www.mobygames.com/game/grand-prix-unlimited", 
"http://www.mobygames.com/game/microsoft-flight-simulator-v30", 
"http://www.mobygames.com/game/have-a-nice-day", 
"http://www.mobygames.com/game/re-volt", 
"http://www.mobygames.com/game/screamer-2", 
"http://www.mobygames.com/game/street-rod", 
"http://www.mobygames.com/game/breakneck", 
"http://www.mobygames.com/game/4x4-off-road-racing", 
"http://www.mobygames.com/game/pitstop-ii", 
"http://www.mobygames.com/game/screamer", 
"http://www.mobygames.com/game/f40-pursuit-simulator", 
"http://www.mobygames.com/game/crazy-cars", 
"http://www.mobygames.com/game/stunt-track-racer", 
"http://www.mobygames.com/game/lamborghini-american-challenge", 
"http://www.mobygames.com/game/superbike-challenge", 
"http://www.mobygames.com/game/vette", 
"http://www.mobygames.com/game/flight-unlimited", 
"http://www.mobygames.com/game/stunt-island", 
"http://www.mobygames.com/game/need-for-speed-high-stakes", 
"http://www.mobygames.com/game/indycar-racing", 
"http://www.mobygames.com/game/ford-simulator", 
"http://www.mobygames.com/game/toca-2-touring-car-challenge", 
"http://www.mobygames.com/game/motocross-madness", 
"http://www.mobygames.com/game/road-hog", 
"http://www.mobygames.com/game/test-drive-off-road", 
"http://www.mobygames.com/game/solo-flight", 
"http://www.mobygames.com/game/ferrari-formula-one", 
"http://www.mobygames.com/game/need-for-speed-porsche-unleashed", 
"http://www.mobygames.com/game/days-of-thunder", 
"http://www.mobygames.com/game/grand-prix-circuit", 
"http://www.mobygames.com/game/american-challenge-a-sailing-simulation", 
"http://www.mobygames.com/game/grand-prix-ii", 
"http://www.mobygames.com/game/nascar-racing", 
"http://www.mobygames.com/game/ford-racing", 
"http://www.mobygames.com/game/orbiter", 
"http://www.mobygames.com/game/nascar-racing-2", 
"http://www.mobygames.com/game/viper-racing", 
"http://www.mobygames.com/game/microsoft-flight-simulator-v20", 
"http://www.mobygames.com/game/nascar-legends", 
"http://www.mobygames.com/game/colin-mcrae-rally", 
"http://www.mobygames.com/game/sega-rally-championship", 
"http://www.mobygames.com/game/real-flight-rc-simulator-deluxe", 
"http://www.mobygames.com/game/cutthroats-terror-on-the-high-seas", 
"http://www.mobygames.com/game/ford-simulator-50", 
"http://www.mobygames.com/game/world-circuit", 
"http://www.mobygames.com/game/microsoft-space-simulator", 
"http://www.mobygames.com/game/microsoft-flight-simulator-v40", 
"http://www.mobygames.com/game/duel-test-drive-ii", 
"http://www.mobygames.com/game/ufo", 
"http://www.mobygames.com/game/lombard-rac-rally", 
"http://www.mobygames.com/game/andretti-racing", 
"http://www.mobygames.com/game/nascar-heat", 
"http://www.mobygames.com/game/grand-prix-3", 
"http://www.mobygames.com/game/f1-racing-simulation", 
"http://www.mobygames.com/game/toca-championship-racing", 
"http://www.mobygames.com/game/chuck-yeagers-advanced-flight-trainer-20", 
"http://www.mobygames.com/game/flight-unlimited-iii", 
"http://www.mobygames.com/game/pro-pilot-99", 
"http://www.mobygames.com/game/midtown-madness-2", 
"http://www.mobygames.com/game/grand-prix-legends", 
"http://www.mobygames.com/game/manic-karts", 
"http://www.mobygames.com/game/lego-racers", 
"http://www.mobygames.com/game/test-drive-4", 
"http://www.mobygames.com/game/dirt-track-racing-sprint-cars", 
"http://www.mobygames.com/game/scars", 
"http://www.mobygames.com/game/speed-busters-american-highways", 
"http://www.mobygames.com/game/flight-assignment-airline-transport-pilot", 
"http://www.mobygames.com/game/screamer-4x4", 
"http://www.mobygames.com/game/test-drive-6", 
"http://www.mobygames.com/game/south-park-rally", 
"http://www.mobygames.com/game/f1-world-grand-prix", 
"http://www.mobygames.com/game/fly-2k", 
"http://www.mobygames.com/game/pocket-rockets", 
"http://www.mobygames.com/game/dirt-track-racing-australia", 
"http://www.mobygames.com/game/ihra-motorsports-drag-racing", 
"http://www.mobygames.com/game/virtual-skipper", 
"http://www.mobygames.com/game/dukes-of-hazzard-racing-for-home", 
"http://www.mobygames.com/game/hard-truck-2", 
"http://www.mobygames.com/game/nascar-racing-3", 
"http://www.mobygames.com/game/nascar-craftsman-truck-series-racing", 
"http://www.mobygames.com/game/beetle-buggin", 
"http://www.mobygames.com/game/le-mans-24-hours", 
"http://www.mobygames.com/game/superkarts", 
"http://www.mobygames.com/game/soda-off-road-racing", 
"http://www.mobygames.com/game/harley-davidson-wheels-of-freedom", 
"http://www.mobygames.com/game/microsoft-flight-simulator-2000", 
"http://www.mobygames.com/game/nascar-racing-4", 
"http://www.mobygames.com/game/x-plane-6", 
"http://www.mobygames.com/game/a320-airbus-edition-europa", 
"http://www.mobygames.com/game/cabelas-4x4-off-road-adventure", 
"http://www.mobygames.com/game/gran-turismo-2", 
"http://www.mobygames.com/game/driver-2", 
"http://www.mobygames.com/game/diddy-kong-racing", 
"http://www.mobygames.com/game/ridge-racer-v", 
"http://www.mobygames.com/game/rage-racer", 
"http://www.mobygames.com/game/r4-ridge-racer-type-4", 
"http://www.mobygames.com/game/smurf-racer", 
"http://www.mobygames.com/game/f1-championship-season-2000", 
"http://www.mobygames.com/game/walt-disney-world-quest-magical-racing-tour", 
"http://www.mobygames.com/game/speed-devils", 
"http://www.mobygames.com/game/star-wars-super-bombad-racing", 
"http://www.mobygames.com/game/motogp", 
"http://www.mobygames.com/game/microsoft-flight-simulator-v10", 
"http://www.mobygames.com/game/abc-sports-indy-racing", 
"http://www.mobygames.com/game/nascar-2001", 
"http://www.mobygames.com/game/mega-man-battle-chase", 
"http://www.mobygames.com/game/grand-prix-500-2", 
"http://www.mobygames.com/game/ace-2", 
"http://www.mobygames.com/game/f1-2000", 
"http://www.mobygames.com/game/harley-davidson-the-road-to-sturgis", 
"http://www.mobygames.com/game/cart-racing", 
"http://www.mobygames.com/game/swedish-touring-car-championship-2", 
"http://www.mobygames.com/game/extreme-500", 
"http://www.mobygames.com/game/burnout-championship-drag-racing", 
"http://www.mobygames.com/game/skunny-kart", 
"http://www.mobygames.com/game/f355-challenge-passione-rossa", 
"http://www.mobygames.com/game/crazy-taxi-2", 
"http://www.mobygames.com/game/b-1-nuclear-bomber", 
"http://www.mobygames.com/game/silkolene-honda-motocross-gp", 
"http://www.mobygames.com/game/pilotwings-64", 
"http://www.mobygames.com/game/formula-one-99", 
"http://www.mobygames.com/game/cyclemania", 
"http://www.mobygames.com/game/chuck-yeagers-advanced-flight-simulator", 
"http://www.mobygames.com/game/hard-drivin", 
"http://www.mobygames.com/game/automobili-lamborghini", 
"http://www.mobygames.com/game/nascar-2000", 
"http://www.mobygames.com/game/ace-combat-3-electrosphere", 
"http://www.mobygames.com/game/ihra-motorsports", 
"http://www.mobygames.com/game/bill-elliotts-nascar-challenge", 
"http://www.mobygames.com/game/panzer-elite", 
"http://www.mobygames.com/game/gadget-racers", 
"http://www.mobygames.com/game/stunt-gp", 
"http://www.mobygames.com/game/rally-trophy", 
"http://www.mobygames.com/game/flight-unlimited-ii", 
"http://www.mobygames.com/game/hard-drivin-ii", 
"http://www.mobygames.com/game/official-formula-one-racing", 
"http://www.mobygames.com/game/f1-world-grand-prix__", 
"http://www.mobygames.com/game/rallisport-challenge", 
"http://www.mobygames.com/game/harley-davidson-race-across-america", 
"http://www.mobygames.com/game/harley-davidson-race-around-the-world", 
"http://www.mobygames.com/game/microsoft-flight-simulator-2002", 
"http://www.mobygames.com/game/f1-2001", 
"http://www.mobygames.com/game/race-drivin", 
"http://www.mobygames.com/game/turbo_", 
"http://www.mobygames.com/game/sega-gt", 
"http://www.mobygames.com/game/star-wars-racer-revenge", 
"http://www.mobygames.com/game/trainz-virtual-railroading-on-your-pc", 
"http://www.mobygames.com/game/taxi-racer", 
"http://www.mobygames.com/game/shrek-swamp-kart-speedway", 
"http://www.mobygames.com/game/italian-job", 
"http://www.mobygames.com/game/chocobo-racing", 
"http://www.mobygames.com/game/excitebike-64", 
"http://www.mobygames.com/game/atv-offroad-fury", 
"http://www.mobygames.com/game/gt-advance-championship-racing", 
"http://www.mobygames.com/game/ridge-racer", 
"http://www.mobygames.com/game/taxi-challenge-berlin", 
"http://www.mobygames.com/game/toyland-racing", 
"http://www.mobygames.com/game/pilotwings", 
"http://www.mobygames.com/game/super-monaco-gp", 
"http://www.mobygames.com/game/checkered-flag", 
"http://www.mobygames.com/game/blue-angels-formation-flight-simulation", 
"http://www.mobygames.com/game/us-racer", 
"http://www.mobygames.com/game/stuntman", 
"http://www.mobygames.com/game/atari-karts", 
"http://www.mobygames.com/game/stunt-driver", 
"http://www.mobygames.com/game/v-rally-edition-99", 
"http://www.mobygames.com/game/need-for-speed-v-rally-2", 
"http://www.mobygames.com/game/crazy-taxi-3-high-roller", 
"http://www.mobygames.com/game/v-rally-3", 
"http://www.mobygames.com/game/world-tour-racing", 
"http://www.mobygames.com/game/gran-turismo-concept-2002-tokyo-geneva", 
"http://www.mobygames.com/game/nascar-racers", 
"http://www.mobygames.com/game/f1roc-race-of-champions", 
"http://www.mobygames.com/game/street-racer", 
"http://www.mobygames.com/game/mafia", 
"http://www.mobygames.com/game/antz-extreme-racing", 
"http://www.mobygames.com/game/motocross-championship", 
"http://www.mobygames.com/game/rad-racer", 
"http://www.mobygames.com/game/grand-prix-4", 
"http://www.mobygames.com/game/grand-prix-challenge", 
"http://www.mobygames.com/game/orbiter__", 
"http://www.mobygames.com/game/swedish-touring-car-championship", 
"http://www.mobygames.com/game/matchbox-caterpillar-construction-zone", 
"http://www.mobygames.com/game/speed-challenge-jacques-villeneuves-racing-vision", 
"http://www.mobygames.com/game/konami-krazy-racers", 
"http://www.mobygames.com/game/hard-truck-road-to-victory", 
"http://www.mobygames.com/game/hard-truck-18-wheels-of-steel", 
"http://www.mobygames.com/game/tyco-rc-assault-with-a-battery", 
"http://www.mobygames.com/game/colin-mcrae-rally-20", 
"http://www.mobygames.com/game/ridge-racer-64", 
"http://www.mobygames.com/game/formula-1", 
"http://www.mobygames.com/game/michael-schumacher-racing-world-kart-2002", 
"http://www.mobygames.com/game/colin-mcrae-rally-3", 
"http://www.mobygames.com/game/motogp-3", 
"http://www.mobygames.com/game/sega-gt-2002", 
"http://www.mobygames.com/game/pro-race-driver", 
"http://www.mobygames.com/game/porsche-challenge", 
"http://www.mobygames.com/game/nascar-dirt-to-daytona", 
"http://www.mobygames.com/game/crazy-taxi-catch-a-ride", 
"http://www.mobygames.com/game/cart-precision-racing", 
"http://www.mobygames.com/game/big-scale-racing", 
"http://www.mobygames.com/game/big-mutha-truckers", 
"http://www.mobygames.com/game/mercedes-benz-truck-racing", 
"http://www.mobygames.com/game/rs3-racing-simulation-three", 
"http://www.mobygames.com/game/rc-helicopter-indoor-flight-simulation", 
"http://www.mobygames.com/game/dirt-track-racing-2", 
"http://www.mobygames.com/game/rc-racers-ii", 
"http://www.mobygames.com/game/gt-advance-2-rally-racing", 
"http://www.mobygames.com/game/gt-advance-3-pro-concept-racing", 
"http://www.mobygames.com/game/muppet-racemania", 
"http://www.mobygames.com/game/f1-pole-position-64", 
"http://www.mobygames.com/game/midtown-madness-3", 
"http://www.mobygames.com/game/a2-racer", 
"http://www.mobygames.com/game/nascar-racing-2003-season", 
"http://www.mobygames.com/game/ford-simulator-iii", 
"http://www.mobygames.com/game/rc-revenge-pro", 
"http://www.mobygames.com/game/hot-wheels-williams-f1-team-racer", 
"http://www.mobygames.com/game/microsoft-flight-simulator-2004-a-century-of-flight", 
"http://www.mobygames.com/game/shuttle-the-space-flight-simulator", 
"http://www.mobygames.com/game/aircars", 
"http://www.mobygames.com/game/world-rally-fever-born-on-the-road", 
"http://www.mobygames.com/game/test-drive-", 
"http://www.mobygames.com/game/king-of-route-66", 
"http://www.mobygames.com/game/motogp-2", 
"http://www.mobygames.com/game/italian-job-", 
"http://www.mobygames.com/game/rendezvous-a-space-shuttle-flight-simulation", 
"http://www.mobygames.com/game/dukes-of-hazzard", 
"http://www.mobygames.com/game/dukes-of-hazzard-ii-daisy-dukes-it-out", 
"http://www.mobygames.com/game/power-drift", 
"http://www.mobygames.com/game/nascar-thunder-2004", 
"http://www.mobygames.com/game/london-racer-world-challenge", 
"http://www.mobygames.com/game/18-wheels-of-steel-across-america", 
"http://www.mobygames.com/game/gran-turismo-concept-2001-tokyo", 
"http://www.mobygames.com/game/colin-mcrae-rally-04", 
"http://www.mobygames.com/game/coronel-indoor-kartracing", 
"http://www.mobygames.com/game/speed-buggy", 
"http://www.mobygames.com/game/aerowings", 
"http://www.mobygames.com/game/nascar-99", 
"http://www.mobygames.com/game/atv-offroad-fury-2", 
"http://www.mobygames.com/game/hot-wheels-stunt-track-driver-2-get-n-dirty", 
"http://www.mobygames.com/game/indycar-series", 
"http://www.mobygames.com/game/f1-world-championship-edition", 
"http://www.mobygames.com/game/need-for-speed-underground", 
"http://www.mobygames.com/game/nascar-thunder-2003", 
"http://www.mobygames.com/game/fatal-run", 
"http://www.mobygames.com/game/speed-punks", 
"http://www.mobygames.com/game/formula-1-championship-edition", 
"http://www.mobygames.com/game/test-drive-off-road-wide-open", 
"http://www.mobygames.com/game/space-shuttle-a-journey-into-space", 
"http://www.mobygames.com/game/age-of-sail-ii-privateers-bounty", 
"http://www.mobygames.com/game/sega-gt-online", 
"http://www.mobygames.com/game/formula-one-2003", 
"http://www.mobygames.com/game/looney-tunes-space-race", 
"http://www.mobygames.com/game/cart-world-series", 
"http://www.mobygames.com/game/rc-cars", 
"http://www.mobygames.com/game/jarrett-labonte-stock-car-racing", 
"http://www.mobygames.com/game/formula-one-world-championship-beyond-the-limit", 
"http://www.mobygames.com/game/test-drive-off-road-2", 
"http://www.mobygames.com/game/atv-quad-power-racing-2", 
"http://www.mobygames.com/game/great-escape-", 
"http://www.mobygames.com/game/road-trip-shifting-gears", 
"http://www.mobygames.com/game/vroom", 
"http://www.mobygames.com/game/toca-race-driver-2", 
"http://www.mobygames.com/game/construction-destruction", 
"http://www.mobygames.com/game/no-second-prize", 
"http://www.mobygames.com/game/revs", 
"http://www.mobygames.com/game/test-drive-le-mans", 
"http://www.mobygames.com/game/nascar-heat-2002", 
"http://www.mobygames.com/game/xtreme-racing", 
"http://www.mobygames.com/game/twisted-metal-small-brawl", 
"http://www.mobygames.com/game/f1-career-challenge", 
"http://www.mobygames.com/game/rc-revenge", 
"http://www.mobygames.com/game/formula-karts", 
"http://www.mobygames.com/game/virtual-karting", 
"http://www.mobygames.com/game/indy-racing-2000", 
"http://www.mobygames.com/game/flight-simulator-ii", 
"http://www.mobygames.com/game/crisis-team-ambulance-driver", 
"http://www.mobygames.com/game/super-cycle", 
"http://www.mobygames.com/game/sito-pons-500-cc-grand-prix", 
"http://www.mobygames.com/game/looney-tunes-racing", 
"http://www.mobygames.com/game/angel-nieto-pole-500", 
"http://www.mobygames.com/game/ihra-drag-racing-2004", 
"http://www.mobygames.com/game/4x4-evo-2", 
"http://www.mobygames.com/game/test-drive-eve-of-destruction", 
"http://www.mobygames.com/game/stunt-race-fx", 
"http://www.mobygames.com/game/nigel-mansells-world-championship-racing", 
"http://www.mobygames.com/game/full-throttle_", 
"http://www.mobygames.com/game/18-wheels-of-steel-pedal-to-the-metal", 
"http://www.mobygames.com/game/ford-racing-2", 
"http://www.mobygames.com/game/warm-up", 
"http://www.mobygames.com/game/turbo-esprit", 
"http://www.mobygames.com/game/combo-racer", 
"http://www.mobygames.com/game/gran-turismo-4-prologue", 
"http://www.mobygames.com/game/power-f1", 
"http://www.mobygames.com/game/rc-stunt-copter", 
"http://www.mobygames.com/game/moonshine-racers", 
"http://www.mobygames.com/game/richard-burns-rally", 
"http://www.mobygames.com/game/rallisport-challenge-2", 
"http://www.mobygames.com/game/colin-mcrae-rally-2005", 
"http://www.mobygames.com/game/touring-car-racer", 
"http://www.mobygames.com/game/need-for-speed-underground-2", 
"http://www.mobygames.com/game/digimon-racing", 
"http://www.mobygames.com/game/motocross-racer", 
"http://www.mobygames.com/game/formula-one_", 
"http://www.mobygames.com/game/ridge-racer-ds", 
"http://www.mobygames.com/game/jetset", 
"http://www.mobygames.com/game/chequered-flag", 
"http://www.mobygames.com/game/rvf-honda", 
"http://www.mobygames.com/game/750cc-grand-prix", 
"http://www.mobygames.com/game/johnny-herberts-grand-prix-championship-1998", 
"http://www.mobygames.com/game/asphalt-urban-gt", 
"http://www.mobygames.com/game/airliner", 
"http://www.mobygames.com/game/radikal-bikers", 
"http://www.mobygames.com/game/f1-world-grand-prix_", 
"http://www.mobygames.com/game/nigel-mansells-grand-prix", 
"http://www.mobygames.com/game/formula-one-2002", 
"http://www.mobygames.com/game/test-drive-2001", 
"http://www.mobygames.com/game/gran-turismo-4", 
"http://www.mobygames.com/game/ferrari-grand-prix-challenge_", 
"http://www.mobygames.com/game/alfa-romeo-racing-italiano", 
"http://www.mobygames.com/game/top-gear", 
"http://www.mobygames.com/game/flight-simulator", 
"http://www.mobygames.com/game/ford-racing-3", 
"http://www.mobygames.com/game/test-drive-off-road-3", 
"http://www.mobygames.com/game/rad-racer-ii", 
"http://www.mobygames.com/game/juiced", 
"http://www.mobygames.com/game/woody-woodpecker-racing", 
"http://www.mobygames.com/game/big-mutha-truckers-2", 
"http://www.mobygames.com/game/formula-one-05", 
"http://www.mobygames.com/game/nascar-2005-chase-for-the-cup", 
"http://www.mobygames.com/game/500-cc-grand-prix", 
"http://www.mobygames.com/game/dukes-of-hazzard-return-of-the-general-lee", 
"http://www.mobygames.com/game/formula-one-04", 
"http://www.mobygames.com/game/choro-q_", 
"http://www.mobygames.com/game/wreckin-crew", 
"http://www.mobygames.com/game/flight-sim-toolkit", 
"http://www.mobygames.com/game/king-of-the-road", 
"http://www.mobygames.com/game/pitstop", 
"http://www.mobygames.com/game/richard-pettys-talladega", 
"http://www.mobygames.com/game/ace-combat-5-the-unsung-war", 
"http://www.mobygames.com/game/formula-1-simulator", 
"http://www.mobygames.com/game/need-for-speed-most-wanted", 
"http://www.mobygames.com/game/acrojet", 
"http://www.mobygames.com/game/ridge-racer_", 
"http://www.mobygames.com/game/newmanhaas-indycar-featuring-nigel-mansell", 
"http://www.mobygames.com/game/agile-warrior-f-111x", 
"http://www.mobygames.com/game/prost-grand-prix-1998", 
"http://www.mobygames.com/game/digital-road", 
"http://www.mobygames.com/game/stuntcar-extreme", 
"http://www.mobygames.com/game/stunt-run", 
"http://www.mobygames.com/game/stuntcar-extreme-advanced", 
"http://www.mobygames.com/game/ridge-racer-6", 
"http://www.mobygames.com/game/moorhuhn-kart-xxl", 
"http://www.mobygames.com/game/crazy-chicken-kart-extra", 
"http://www.mobygames.com/game/crazy-chicken-kart-2", 
"http://www.mobygames.com/game/road-trip-the-arcade-edition", 
"http://www.mobygames.com/game/atv-mudracer", 
"http://www.mobygames.com/game/amsterdam-taxi-madness", 
"http://www.mobygames.com/game/mario-andretti-racing", 
"http://www.mobygames.com/game/motogp_", 
"http://www.mobygames.com/game/f1-grand-prix", 
"http://www.mobygames.com/game/drivers-education-98", 
"http://www.mobygames.com/game/speed-king-2", 
"http://www.mobygames.com/game/sonic-riders", 
"http://www.mobygames.com/game/f-1-world-grand-prix-ii", 
"http://www.mobygames.com/game/crashday", 
"http://www.mobygames.com/game/f-1-world-grand-prix", 
"http://www.mobygames.com/game/sega-rally-championship_", 
"http://www.mobygames.com/game/revs_", 
"http://www.mobygames.com/game/mad-tracks", 
"http://www.mobygames.com/game/cart-flag-to-flag", 
"http://www.mobygames.com/game/drivin-force", 
"http://www.mobygames.com/game/asphalt-urban-gt_", 
"http://www.mobygames.com/game/v-rally-2", 
"http://www.mobygames.com/game/speed-devils_", 
"http://www.mobygames.com/game/microsoft-flight-simulator", 
"http://www.mobygames.com/game/speedboat-assassins", 
"http://www.mobygames.com/game/hot-wheels-stunt-track-driver", 
"http://www.mobygames.com/game/world-racing-2", 
"http://www.mobygames.com/game/nascar-06-total-team-control", 
"http://www.mobygames.com/game/great-american-cross-country-road-race", 
"http://www.mobygames.com/game/motogp-06", 
"http://www.mobygames.com/game/need-for-speed-underground-2_", 
"http://www.mobygames.com/game/nascar-98", 
"http://www.mobygames.com/game/flyin-high", 
"http://www.mobygames.com/game/mickeys-speedway-usa-", 
"http://www.mobygames.com/game/big-rigs-over-the-road-racing", 
"http://www.mobygames.com/game/formula-one-2000", 
"http://www.mobygames.com/game/s40-racing", 
"http://www.mobygames.com/game/steel-talons", 
"http://www.mobygames.com/game/toca-race-driver-3", 
"http://www.mobygames.com/game/corvette-evolution-gt", 
"http://www.mobygames.com/game/rally-challenge", 
"http://www.mobygames.com/game/atv-quad-frenzy", 
"http://www.mobygames.com/game/shuttle-orbiter", 
"http://www.mobygames.com/game/fly-ii", 
"http://www.mobygames.com/game/x-plane-8", 
"http://www.mobygames.com/game/desert-racing-of-bardos", 
"http://www.mobygames.com/game/gtr-2-fia-gt-racing-game", 
"http://www.mobygames.com/game/bmw-williamsf1-team-racing-challenge", 
"http://www.mobygames.com/game/flight-light-plus", 
"http://www.mobygames.com/game/super-1-karting-simulation", 
"http://www.mobygames.com/game/american-chopper", 
"http://www.mobygames.com/game/ford-vs-chevy", 
"http://www.mobygames.com/game/need-for-speed-carbon", 
"http://www.mobygames.com/game/advan-racing", 
"http://www.mobygames.com/game/excite-truck", 
"http://www.mobygames.com/game/gt-pro-series", 
"http://www.mobygames.com/game/need-for-speed-most-wanted-5-1-0", 
"http://www.mobygames.com/game/american-chopper-2-full-throttle", 
"http://www.mobygames.com/game/nascar-simracing", 
"http://www.mobygames.com/game/ship-simulator-2006-collectors-edition", 
"http://www.mobygames.com/game/asphalt-3-street-rules", 
"http://www.mobygames.com/game/paragliding", 
"http://www.mobygames.com/game/need-for-speed-carbon-own-the-city", 
"http://www.mobygames.com/game/aircraft-powerpack", 
"http://www.mobygames.com/game/asphalt-urban-gt-2", 
"http://www.mobygames.com/game/microsoft-flight-simulator-x", 
"http://www.mobygames.com/game/race-the-official-wtcc-game", 
"http://www.mobygames.com/game/cartoon-network-racing", 
"http://www.mobygames.com/game/colin-mcrae-rally-2005-", 
"http://www.mobygames.com/game/toca-race-driver-3-challenge", 
"http://www.mobygames.com/game/diddy-kong-racing-ds", 
"http://www.mobygames.com/game/ridge-racer-7", 
"http://www.mobygames.com/game/prime-mover", 
"http://www.mobygames.com/game/test-drive-unlimited", 
"http://www.mobygames.com/game/motogp-ultimate-racing-technology-3", 
"http://www.mobygames.com/game/formula-1-championship-edition_", 
"http://www.mobygames.com/game/cyberbykes-shadow-racer-vr", 
"http://www.mobygames.com/game/wheels-on-fire", 
"http://www.mobygames.com/game/paris-marseille-racing", 
"http://www.mobygames.com/game/europe-racing", 
"http://www.mobygames.com/game/wing-island", 
"http://www.mobygames.com/game/pac-man-world-rally", 
"http://www.mobygames.com/game/mario-kart-double-dash-special-edition", 
"http://www.mobygames.com/game/mario-andrettis-racing-challenge", 
"http://www.mobygames.com/game/xcite-bike", 
"http://www.mobygames.com/game/motorstorm", 
"http://www.mobygames.com/game/rayman-kart", 
"http://www.mobygames.com/game/ford-simulator-ii", 
"http://www.mobygames.com/game/ford-bold-moves-street-racing", 
"http://www.mobygames.com/game/sabans-power-rangers-zeo-battle-racers", 
"http://www.mobygames.com/game/touring-car-champions", 
"http://www.mobygames.com/game/dirt", 
"http://www.mobygames.com/game/off-road-challenge", 
"http://www.mobygames.com/game/motogp-07", 
"http://www.mobygames.com/game/sega-rally-revo", 
"http://www.mobygames.com/game/stuntman-ignition", 
"http://www.mobygames.com/game/race-07-official-wtcc-game", 
"http://www.mobygames.com/game/crazy-taxi-fare-wars", 
"http://www.mobygames.com/game/need-for-speed-prostreet", 
"http://www.mobygames.com/game/nascar-08", 
"http://www.mobygames.com/game/atv-thunder-ridge-riders-monster-trucks-mayhem", 
"http://www.mobygames.com/game/ford-racing-3_", 
"http://www.mobygames.com/game/super-bike-transam", 
"http://www.mobygames.com/game/twin-turbo-v8", 
"http://www.mobygames.com/game/team-losi-rc-racer", 
"http://www.mobygames.com/game/mms-kart-racing", 
"http://www.mobygames.com/game/bill-elliotts-nascar-fast-tracks", 
"http://www.mobygames.com/game/nascar-98-collectors-edition", 
"http://www.mobygames.com/game/african-raiders-01", 
"http://www.mobygames.com/game/turbo-cup", 
"http://www.mobygames.com/game/burnout-dominator", 
"http://www.mobygames.com/game/motogp-07_", 
"http://www.mobygames.com/game/parigi-dakar", 
"http://www.mobygames.com/game/2-games-toca-world-touring-cars-colin-mcrae-rally", 
"http://www.mobygames.com/game/truckin", 
"http://www.mobygames.com/game/red-arrows", 
"http://www.mobygames.com/game/formula-one-arcade", 
"http://www.mobygames.com/game/500cc-motomanager", 
"http://www.mobygames.com/game/london-racer-ii", 
"http://www.mobygames.com/game/mms-kart-racing_", 
"http://www.mobygames.com/game/gran-turismo-5-prologue", 
"http://www.mobygames.com/game/f1-manager__", 
"http://www.mobygames.com/game/sega-rally-revo_", 
"http://www.mobygames.com/game/grid", 
"http://www.mobygames.com/game/gran-turismo-hd-concept", 
"http://www.mobygames.com/game/road-trip", 
"http://www.mobygames.com/game/sonic-riders-zero-gravity", 
"http://www.mobygames.com/game/need-for-speed-most-wanted_", 
"http://www.mobygames.com/game/atv-offroad-fury-blazin-trails", 
"http://www.mobygames.com/game/ferrari-grand-prix-challenge", 
"http://www.mobygames.com/game/shuttle-simulator", 
"http://www.mobygames.com/game/virtual-skipper-2", 
"http://www.mobygames.com/game/night-racer", 
"http://www.mobygames.com/game/formula-1-3d", 
"http://www.mobygames.com/game/sonic-drift-2", 
"http://www.mobygames.com/game/dirt-trax-fx", 
"http://www.mobygames.com/game/indy-500_", 
"http://www.mobygames.com/game/vmx-racing", 
"http://www.mobygames.com/game/ayrton-senna-kart-duel-2", 
"http://www.mobygames.com/game/super-trucks-racing", 
"http://www.mobygames.com/game/a2-racer-ii", 
"http://www.mobygames.com/game/a2-racer-iii-europa-tour", 
"http://www.mobygames.com/game/a2-racer-iv-the-cops-revenge", 
"http://www.mobygames.com/game/london-racer-police-madness", 
"http://www.mobygames.com/game/london-racer-destruction-madness", 
"http://www.mobygames.com/game/autobahn-raser", 
"http://www.mobygames.com/game/autobahn-raser-ii", 
"http://www.mobygames.com/game/nascar-revolution", 
"http://www.mobygames.com/game/nascar-revolution-se", 
"http://www.mobygames.com/game/ferrari-world-championship", 
"http://www.mobygames.com/game/nascar-09", 
"http://www.mobygames.com/game/f-1-world-grand-prix_", 
"http://www.mobygames.com/game/big-rig-europe", 
"http://www.mobygames.com/game/final-lap-2000", 
"http://www.mobygames.com/game/f1-world-grand-prix-ii", 
"http://www.mobygames.com/game/ship-simulator-2008", 
"http://www.mobygames.com/game/ship-simulator-2006-professional", 
"http://www.mobygames.com/game/test-drive-le-mans_", 
"http://www.mobygames.com/game/motogp-08", 
"http://www.mobygames.com/game/indianapolis-500-legends", 
"http://www.mobygames.com/game/ferrari-challenge-trofeo-pirelli", 
"http://www.mobygames.com/game/need-for-speed-undercover", 
"http://www.mobygames.com/game/looney-tunes-racing_", 
"http://www.mobygames.com/game/nira-intense-import-drag-racing", 
"http://www.mobygames.com/game/atv-offroad-fury-3", 
"http://www.mobygames.com/game/f1-world-grand-prix-ii-for-game-boy-color", 
"http://www.mobygames.com/game/matchbox-caterpillar-big-dirt-movers", 
"http://www.mobygames.com/game/complete-simbin-pack", 
"http://www.mobygames.com/game/penny-racers", 
"http://www.mobygames.com/game/burning-road", 
"http://www.mobygames.com/game/ford-drive-gold", 
"http://www.mobygames.com/game/subway-simulator-volume-1-the-path-new-york-underground", 
"http://www.mobygames.com/game/quad", 
"http://www.mobygames.com/game/rfactor", 
"http://www.mobygames.com/game/virtual-skipper-3", 
"http://www.mobygames.com/game/virtual-skipper-4", 
"http://www.mobygames.com/game/virtual-skipper-5-32nd-americas-cup-the-game", 
"http://www.mobygames.com/game/virtual-skipper-online", 
"http://www.mobygames.com/game/virtual-skipper-5", 
"http://www.mobygames.com/game/motogp__", 
"http://www.mobygames.com/game/motogp-4", 
"http://www.mobygames.com/game/motogp-2_", 
"http://www.mobygames.com/game/evel-knievel-interactive-stunt-game", 
"http://www.mobygames.com/game/gp-1-part-ii", 
"http://www.mobygames.com/game/ship-simulator-2006", 
"http://www.mobygames.com/game/18-wheels-of-steel-haulin", 
"http://www.mobygames.com/game/18-wheels-of-steel-convoy", 
"http://www.mobygames.com/game/motogp-ultimate-racing-technology", 
"http://www.mobygames.com/game/international-moto-x", 
"http://www.mobygames.com/game/harley-davidson-race-to-the-rally", 
"http://www.mobygames.com/game/volvo-the-game", 
"http://www.mobygames.com/game/asphalt-4-elite-racing", 
"http://www.mobygames.com/game/ridge-racer-2", 
"http://www.mobygames.com/game/pilot-academy", 
"http://www.mobygames.com/game/nascar-racing-2002-season", 
"http://www.mobygames.com/game/microsoft-flight-simulator-2000-professional-edition", 
"http://www.mobygames.com/game/skidoo", 
"http://www.mobygames.com/game/urlaubs-raser", 
"http://www.mobygames.com/game/fly-2k-german-edition", 
"http://www.mobygames.com/game/open-kart", 
"http://www.mobygames.com/game/microsoft-flight-simulator-2002-professional-edition", 
"http://www.mobygames.com/game/need-for-speed-shift", 
"http://www.mobygames.com/game/autobahn-raser-das-spiel-zum-film", 
"http://www.mobygames.com/game/rc-helicopter", 
"http://www.mobygames.com/game/michael-schumacher-world-tour-kart-2004", 
"http://www.mobygames.com/game/taxi-raser", 
"http://www.mobygames.com/game/nicktoons-racing", 
"http://www.mobygames.com/game/formula-challenge", 
"http://www.mobygames.com/game/f1-racing-championship_", 
"http://www.mobygames.com/game/dirt-2", 
"http://www.mobygames.com/game/colin-mcrae-dirt-2-special-edition", 
"http://www.mobygames.com/game/motorace-usa", 
"http://www.mobygames.com/game/need-for-speed-carbon-own-the-city_", 
"http://www.mobygames.com/game/f1-pole-position", 
"http://www.mobygames.com/game/18-wheels-of-steel-american-long-haul", 
"http://www.mobygames.com/game/x-plane-9", 
"http://www.mobygames.com/game/nice-2-king-size", 
"http://www.mobygames.com/game/arca-sim-racing-08", 
"http://www.mobygames.com/game/a320-airbus-edition-usa", 
"http://www.mobygames.com/game/pickup-express", 
"http://www.mobygames.com/game/motogp-0910", 
"http://www.mobygames.com/game/4x4-team", 
"http://www.mobygames.com/game/gran-turismo_", 
"http://www.mobygames.com/game/revs__", 
"http://www.mobygames.com/game/overdrive_", 
"http://www.mobygames.com/game/crazee-rider", 
"http://www.mobygames.com/game/4x4-hummer", 
"http://www.mobygames.com/game/superstars-v8-racing", 
"http://www.mobygames.com/game/bus-simulator-2009", 
"http://www.mobygames.com/game/f-1", 
"http://www.mobygames.com/game/ihra-professional-drag-racing-2005", 
"http://www.mobygames.com/game/karttoon-rally", 
"http://www.mobygames.com/game/18-wheels-of-steel-extreme-trucker", 
"http://www.mobygames.com/game/german-truck-simulator", 
"http://www.mobygames.com/game/superstars-v8-next-challenge", 
"http://www.mobygames.com/game/modnation-racers", 
"http://www.mobygames.com/game/rignroll", 
"http://www.mobygames.com/game/need-for-speed-world", 
"http://www.mobygames.com/game/need-for-speed-prostreet_", 
"http://www.mobygames.com/game/colin-mcrae-rally-2005-plus", 
"http://www.mobygames.com/game/explosive-racing", 
"http://www.mobygames.com/game/indycar-series-2005", 
"http://www.mobygames.com/game/volvo-drive-for-life", 
"http://www.mobygames.com/game/london-cab-challenge", 
"http://www.mobygames.com/game/formula-nippon", 
"http://www.mobygames.com/game/sonic-free-riders", 
"http://www.mobygames.com/game/need-for-speed-nitro-x", 
"http://www.mobygames.com/game/gran-turismo-5-collectors-edition", 
"http://www.mobygames.com/game/gran-turismo-5", 
"http://www.mobygames.com/game/polnyj-privod-ves-uaz-zolotoe-izdanie", 
"http://www.mobygames.com/game/all-star-racing-2", 
"http://www.mobygames.com/game/top-fuel-eliminator", 
"http://www.mobygames.com/game/tonka-monster-trucks", 
"http://www.mobygames.com/game/paris-marseille-racing-ii", 
"http://www.mobygames.com/game/indoor-kart-racing", 
"http://www.mobygames.com/game/asphalt-5", 
"http://www.mobygames.com/game/polnyj-privod-2-hummer-extreme-edition", 
"http://www.mobygames.com/game/slot-car-racer", 
"http://www.mobygames.com/game/pilotwings-resort", 
"http://www.mobygames.com/game/atv-gp", 
"http://www.mobygames.com/game/autobahn-tokio", 
"http://www.mobygames.com/game/nascar-racing_", 
"http://www.mobygames.com/game/excitebots-trick-racing", 
"http://www.mobygames.com/game/shift-2-unleashed", 
"http://www.mobygames.com/game/super-taxi-driver", 
"http://www.mobygames.com/game/taxi-challenge-london", 
"http://www.mobygames.com/game/taxi-challenge-new-york", 
"http://www.mobygames.com/game/f1-2010", 
"http://www.mobygames.com/game/ship-simulator-extremes", 
"http://www.mobygames.com/game/dirt-3", 
"http://www.mobygames.com/game/f-sim-space-shuttle", 
"http://www.mobygames.com/game/toca-world-touring-cars", 
"http://www.mobygames.com/game/sonic-sega-all-stars-racing_", 
"http://www.mobygames.com/game/gti-club-rally-cte-dazur", 
"http://www.mobygames.com/game/nascar-unleashed", 
"http://www.mobygames.com/game/super-taxi-driver-the-original", 
"http://www.mobygames.com/game/need-for-speed-the-run", 
"http://www.mobygames.com/game/f1-2009", 
"http://www.mobygames.com/game/gran-turismo-5-signature-edition", 
"http://www.mobygames.com/game/f1-2011", 
"http://www.mobygames.com/game/tow-truck-simulator-2010", 
"http://www.mobygames.com/game/knight-rider-special", 
"http://www.mobygames.com/game/krazy-kart-racing", 
"http://www.mobygames.com/game/need-for-speed-undercover_", 
"http://www.mobygames.com/game/uk-truck-simulator", 
"http://www.mobygames.com/game/concorde-landing-simulation", 
"http://www.mobygames.com/game/pac-man-kart-rally-3d", 
"http://www.mobygames.com/game/f1-2002_", 
"http://www.mobygames.com/game/dirt-3-complete-edition", 
"http://www.mobygames.com/game/ridge-racer-3d", 
"http://www.mobygames.com/game/microsoft-flight", 
"http://www.mobygames.com/game/ridge-racer-unbounded", 
"http://www.mobygames.com/game/approach-trainer", 
"http://www.mobygames.com/game/ifr", 
"http://www.mobygames.com/game/747-flight-simulator", 
"http://www.mobygames.com/game/flight-path-737", 
"http://www.mobygames.com/game/taxi-3", 
"http://www.mobygames.com/game/international-speedway", 
"http://www.mobygames.com/game/jumbo-lander", 
"http://www.mobygames.com/game/kart-race", 
"http://www.mobygames.com/game/dirt-showdown", 
"http://www.mobygames.com/game/race-drivin_", 
"http://www.mobygames.com/game/pilot", 
"http://www.mobygames.com/game/wroom", 
"http://www.mobygames.com/game/test-drive-ferrari-racing-legends", 
"http://www.mobygames.com/game/need-for-speed-the-run-limited-edition", 
"http://www.mobygames.com/game/need-for-speed-the-run-limited-edition-origincom-pre-order-versi", 
"http://www.mobygames.com/game/turbo-kart-racer_", 
"http://www.mobygames.com/game/turbo-kart-racer", 
"http://www.mobygames.com/game/professional-go-kart-simulator", 
"http://www.mobygames.com/game/gunship__", 
"http://www.mobygames.com/game/tow-things-on-wheels", 
"http://www.mobygames.com/game/saturn-navigator", 
"http://www.mobygames.com/game/xs-moto", 
"http://www.mobygames.com/game/truck-simulator-complete-edition", 
"http://www.mobygames.com/game/transportando-o-brasil", 
"http://www.mobygames.com/game/driving-simulator-2012", 
"http://www.mobygames.com/game/driving-simulator-2011", 
"http://www.mobygames.com/game/f1-2011_", 
"http://www.mobygames.com/game/x-plane-5", 
"http://www.mobygames.com/game/x-plane-version-7", 
"http://www.mobygames.com/game/sierra-pro-pilot-98-the-complete-flight-simulator", 
"http://www.mobygames.com/game/f1-2012", 
"http://www.mobygames.com/game/drag-race-eliminator", 
"http://www.mobygames.com/game/military-collection-2008", 
"http://www.mobygames.com/game/worlds-of-flight", 
"http://www.mobygames.com/game/ultimate-flight-collection", 
"http://www.mobygames.com/game/need-for-speed-most-wanted__", 
"http://www.mobygames.com/game/stunt-jumper", 
"http://www.mobygames.com/game/f1-race-stars", 
"http://www.mobygames.com/game/indy-500__", 
"http://www.mobygames.com/game/x-plane-10-regional-edition-north-america", 
"http://www.mobygames.com/game/sonic-all-stars-racing-transformed", 
"http://www.mobygames.com/game/x-plane-10-global-edition-aerosoft-edition", 
"http://www.mobygames.com/game/a320-airbus-vol-2", 
"http://www.mobygames.com/game/motogp-1011", 
"http://www.mobygames.com/game/driving-simulator-2009", 
"http://www.mobygames.com/game/special-edition-fly-2k", 
"http://www.mobygames.com/game/x-plane-10-global-edition", 
"http://www.mobygames.com/game/grid-2", 
"http://www.mobygames.com/game/f1-circus-cd", 
"http://www.mobygames.com/game/ayrton-senna-kart-duel", 
"http://www.mobygames.com/game/colin-mcrae-rally-20_", 
"http://www.mobygames.com/game/need-for-speed-underground_", 
"http://www.mobygames.com/game/rally-sport", 
"http://www.mobygames.com/game/gran-turismo-6", 
"http://www.mobygames.com/game/euro-truck-simulator-2", 
"http://www.mobygames.com/game/colin-mcrae-rally-2005_", 
"http://www.mobygames.com/game/need-for-speed-rivals", 
"http://www.mobygames.com/game/747_", 
"http://www.mobygames.com/game/glider-pilot", 
"http://www.mobygames.com/game/suzuka-8-hours_", 
"http://www.mobygames.com/game/asphalt-6-adrenaline", 
"http://www.mobygames.com/game/asphalt-7-heat", 
"http://www.mobygames.com/game/f1-pole-position_", 
"http://www.mobygames.com/game/maxx-trucks", 
"http://www.mobygames.com/game/touge-king-the-spirits-2", 
"http://www.mobygames.com/game/dtm-experience", 
"http://www.mobygames.com/game/sega-rally-2006", 
"http://www.mobygames.com/game/f1-2013", 
"http://www.mobygames.com/game/speed-king", 
"http://www.mobygames.com/game/solo-flight-enhanced", 
"http://www.mobygames.com/game/solo-flight-2nd-edition", 
"http://www.mobygames.com/game/nascar-14", 
"http://www.mobygames.com/game/final-frontier_", 
"http://www.mobygames.com/game/mario-kart-arcade-gp", 
"http://www.mobygames.com/game/need-for-speed-nitro", 
"http://www.mobygames.com/game/chevy-chase", 
"http://www.mobygames.com/game/max-torque", 
"http://www.mobygames.com/game/human-grand-prix-iv-f1-dream-battle", 
"http://www.mobygames.com/game/final-stretch", 
"http://www.mobygames.com/game/colin-mcrae-rally_", 
"http://www.mobygames.com/game/motogp-13", 
"http://www.mobygames.com/game/bygg-btar-med-mulle-meck", 
"http://www.mobygames.com/game/mario-kart-arcade-gp-2", 
"http://www.mobygames.com/game/railsimulatorcoms-train-simulator-2013", 
"http://www.mobygames.com/game/trucks-trailers", 
"http://www.mobygames.com/game/f1-race-stars_", 
"http://www.mobygames.com/game/747-400-precision-simulator", 
"http://www.mobygames.com/game/jumbo-jet-pilot", 
"http://www.mobygames.com/game/gran-turismo-5-academy-edition", 
"http://www.mobygames.com/game/test-drive-6_", 
"http://www.mobygames.com/game/top-speed", 
"http://www.mobygames.com/game/garbage-truck-simulator", 
"http://www.mobygames.com/game/18-wheels-of-steel-extreme-trucker-2", 
"http://www.mobygames.com/game/nascar-arcade", 
"http://www.mobygames.com/game/aerofly-fs", 
"http://www.mobygames.com/game/atv-quad-power-racing_", 
"http://www.mobygames.com/game/antz-extreme-racing_", 
"http://www.mobygames.com/game/need-for-speed-porsche-unleashed_", 
"http://www.mobygames.com/game/driveclub", 
"http://www.mobygames.com/game/asphalt-injection", 
"http://www.mobygames.com/game/atv-offroad-fury-pro", 
"http://www.mobygames.com/game/v-rally-3_", 
"http://www.mobygames.com/game/space-shuttle-mission-2007", 
"http://www.mobygames.com/game/scania-truck-driving-simulator-the-game", 
"http://www.mobygames.com/game/beach-buggy-blitz", 
"http://www.mobygames.com/game/littlebigplanet-karting", 
"http://www.mobygames.com/game/monkey-kart", 
"http://www.mobygames.com/game/goofy-gokarts", 
"http://www.mobygames.com/game/pro-pilot-usa", 
"http://www.mobygames.com/game/air-ranger-rescue-helicopter", 
"http://www.mobygames.com/game/project-cars", 
"http://www.mobygames.com/game/grid-autosport", 
"http://www.mobygames.com/game/action-girlz-racing", 
"http://www.mobygames.com/game/taxi-2", 
"http://www.mobygames.com/game/bus-simulator-2012", 
"http://www.mobygames.com/game/beach-buggy-racing", 
"http://www.mobygames.com/game/gt-racers", 
"http://www.mobygames.com/game/grid-start", 
"http://www.mobygames.com/game/kelloggs-frosties-gr-r-rand-prix", 
"http://www.mobygames.com/game/autobahn-raser-iv", 
"http://www.mobygames.com/game/dirt-2_", 
"http://www.mobygames.com/game/mad-max_", 
"http://www.mobygames.com/game/super-monkey-ball-3d", 
"http://www.mobygames.com/game/zoom_", 
"http://www.mobygames.com/game/wacky-karts", 
"http://www.mobygames.com/game/need-for-speed-no-limits", 
"http://www.mobygames.com/game/speed-devils-online-racing", 
"http://www.mobygames.com/game/need-for-speed_", 
"http://www.mobygames.com/game/road-racer", 
"http://www.mobygames.com/game/delta-wing", 
"http://www.mobygames.com/game/asphalt-nitro", 
"http://www.mobygames.com/game/asphalt-nitro_", 
"http://www.mobygames.com/game/garbage-truck-simulator_", 
"http://www.mobygames.com/game/f1-2015", 
"http://www.mobygames.com/game/tanker-truck-simulator-2011", 
"http://www.mobygames.com/game/delivery-truck-simulator", 
"http://www.mobygames.com/game/cocoto-kart-racer", 
"http://www.mobygames.com/game/moorhuhn-fun-kart-2008", 
"http://www.mobygames.com/game/crazy-chicken-kart-thunder", 
"http://www.mobygames.com/game/american-truck-simulator", 
"http://www.mobygames.com/game/need-for-speed-most-wanted___", 
"http://www.mobygames.com/game/turbo-trucks", 
"http://www.mobygames.com/game/taxi-racer-london-2", 
"http://www.mobygames.com/game/motogp___", 
"http://www.mobygames.com/game/f-1-chequered-flag", 
"http://www.mobygames.com/game/gadget-racers_", 
"http://www.mobygames.com/game/wrc-5", 
"http://www.mobygames.com/game/x-plane-8-deluxe", 
"http://www.mobygames.com/game/dirt-rally-legend-edition", 
"http://www.mobygames.com/game/planets_", 
"http://www.mobygames.com/game/f1-2014", 
"http://www.mobygames.com/game/f1-2013-classic-edition", 
"http://www.mobygames.com/game/championship-rally_", 
"http://www.mobygames.com/game/f1-2016", 
"http://www.mobygames.com/game/truck-racing", 
"http://www.mobygames.com/game/asphalt-xtreme", 
"http://www.mobygames.com/game/f1-game-franchise-bundle", 
"http://www.mobygames.com/game/aerofly-fs-2-flight-simulator", 
"http://www.mobygames.com/game/rig-racer-2", 
"http://www.mobygames.com/game/need-for-speed-undercover__", 
"http://www.mobygames.com/game/garfield-kart", 
"http://www.mobygames.com/game/safari-race", 
"http://www.mobygames.com/game/feuerwehr-simulator-2010", 
"http://www.mobygames.com/game/farming-simulator-17", 
"http://www.mobygames.com/game/farm-machines-championships-2014", 
"http://www.mobygames.com/game/furious-karting", 
"http://www.mobygames.com/game/poppy-kart", 
"http://www.mobygames.com/game/truck-racing-by-renault-trucks", 
"http://www.mobygames.com/game/driveclub-vr", 
"http://www.mobygames.com/game/driveclub-bikes", 
"http://www.mobygames.com/game/fireman-sam-fire-rescue", 
"http://www.mobygames.com/game/ride-em-low", 
"http://www.mobygames.com/game/3d-grand-prix_", 
"http://www.mobygames.com/game/call-of-duty-infinite-warfare-jackal-assault-vr-experience", 
"http://www.mobygames.com/game/dirt-rally", 
"http://www.mobygames.com/game/dirt-rally-vr", 
"http://www.mobygames.com/game/motogp-14", 
"http://www.mobygames.com/game/motogp-15", 
"http://www.mobygames.com/game/valentino-rossi-the-game", 
"http://www.mobygames.com/game/beamngdrive", 
"http://www.mobygames.com/game/excitebike-world-rally", 
"http://www.mobygames.com/game/need-for-speed-the-run_", 
"http://www.mobygames.com/game/homie-rollerz", 
"http://www.mobygames.com/game/heracles-chariot-racing", 
"http://www.mobygames.com/game/recycle-garbage-truck-simulator", 
"http://www.mobygames.com/game/need-for-speed-nitro_", 
"http://www.mobygames.com/game/farm-expert-2016", 
"http://www.mobygames.com/game/train-mechanic-simulator-2017", 
"http://www.mobygames.com/game/helicopter-2015-natural-disasters", 
"http://www.mobygames.com/game/star-wars-racer-arcade", 
"http://www.mobygames.com/game/trucker-2", 
"http://www.mobygames.com/game/trucker", 
"http://www.mobygames.com/game/x-plane-11", 
"http://www.mobygames.com/game/mxgp2-the-official-motocross-videogame", 
"http://www.mobygames.com/game/nascar-heat-evolution", 
"http://www.mobygames.com/game/big-mutha-truckers_", 
"http://www.mobygames.com/game/mxgp-the-official-motocross-videogame", 
"http://www.mobygames.com/game/professional-farmer-2017", 
"http://www.mobygames.com/game/interkosmos", 
"http://www.mobygames.com/game/giant-machines-2017", 
"http://www.mobygames.com/game/need-for-speed-shift_", 
"http://www.mobygames.com/game/infinitum", 
"http://www.mobygames.com/game/crazy-chicken-star-karts", 
"http://www.mobygames.com/game/x-plane-11-aerosoft-edition", 
"http://www.mobygames.com/game/quadcopter-pilot-challenge", 
"http://www.mobygames.com/game/train-simulator", 
"http://www.mobygames.com/game/bus-simulator-16", 
"http://www.mobygames.com/game/fireman-sam-junior-cadet", 
"http://www.mobygames.com/game/mechoecho", 
"http://www.mobygames.com/game/flug-simulator-fr-airbus-a-320-europa-edition", 
"http://www.mobygames.com/game/lunar", 
"http://www.mobygames.com/game/rocket_", 
"http://www.mobygames.com/game/lem_", 
"http://www.mobygames.com/game/moonlander", 
"http://www.mobygames.com/game/farming-simulator-18", 
"http://www.mobygames.com/game/red-bull-air-race-the-game", 
"http://www.mobygames.com/game/motogp-17", 
"http://www.mobygames.com/game/aerofly-rc-flight-simulator-rc-7", 
"http://www.mobygames.com/game/aerofly-rc-flight-simulator-rc-7-professional-edition", 
"http://www.mobygames.com/game/aerofly-rc-flight-simulator-rc-7-ultimate-edition", 
"http://www.mobygames.com/game/island-flight-simulator", 
"http://www.mobygames.com/game/macross-delta-scramble", 
"http://www.mobygames.com/game/omsi-2", 
"http://www.mobygames.com/game/citybus-simulator-munich", 
"http://www.mobygames.com/game/dive-to-the-titanic", 
"http://www.mobygames.com/game/world-of-subways-4-subway-simulator-new-york-queens-manhattan-li", 
"http://www.mobygames.com/game/on-the-road-truck-simulator", 
"http://www.mobygames.com/game/glider-island", 
"http://www.mobygames.com/game/a-train-3d-city-simulator", 
"http://www.mobygames.com/game/fsw-flight-sim-world", 
"http://www.mobygames.com/game/american-powerhaul-train-simulator", 
"http://www.mobygames.com/game/tgv-voyages-train-simulator", 
"http://www.mobygames.com/game/tower-3d", 
"http://www.mobygames.com/game/tower-3d-pro", 
"http://www.mobygames.com/game/tower-2011se", 
"http://www.mobygames.com/game/boat-fury", 
"http://www.mobygames.com/game/moto-cross", 
"http://www.mobygames.com/game/f1-race-stars-powered-up-edition", 
"http://www.mobygames.com/game/micro-flight", 
"http://www.mobygames.com/game/f1-2016_", 
"http://www.mobygames.com/game/trial-challenge-world-tour", 
"http://www.mobygames.com/game/eep-12-train-simulator-expert", 
"http://www.mobygames.com/game/eep-13-eisenbahnsimulation", 
"http://www.mobygames.com/game/eep-train-simulator-mission", 
"http://www.mobygames.com/game/vrc-pro", 
"http://www.mobygames.com/game/truck-simulator-pro-2", 
"http://www.mobygames.com/game/infinite-flight", 
"http://www.mobygames.com/game/automation", 
"http://www.mobygames.com/game/f1-2017", 
"http://www.mobygames.com/game/formula-e-grand-prix", 
"http://www.mobygames.com/game/formula-vr", 
"http://www.mobygames.com/game/heavyweight-transport-simulator", 
"http://www.mobygames.com/game/off-road-drive", 
"http://www.mobygames.com/game/autobahn_", 
"http://www.mobygames.com/game/nascar-heat-2", 
"http://www.mobygames.com/game/atv-wild-ride-3d", 
"http://www.mobygames.com/game/face-racers-photo-finish", 
"http://www.mobygames.com/game/need-for-speed-underground-2__", 
"http://www.mobygames.com/game/ferrari-challenge-trofeo-pirelli_", 
"http://www.mobygames.com/game/flight-simulator-vr", 
"http://www.mobygames.com/game/off-road-vr", 
"http://www.mobygames.com/game/tracks-the-train-set-game", 
"http://www.mobygames.com/game/gran-turismo-sport", 
"http://www.mobygames.com/game/kart-racer", 
"http://www.mobygames.com/game/hawker-heroes", 
"http://www.mobygames.com/game/flight-deck-6", 
"http://www.mobygames.com/game/18-wheels-of-steel-gold", 
"http://www.mobygames.com/game/18-wheels-of-steel-gold-dvd", 
"http://www.mobygames.com/game/professional-farmer-american-dream", 
"http://www.mobygames.com/game/fly-the-boeing-787-dreamliner", 
"http://www.mobygames.com/game/cars-in-the-fast-lane", 
"http://www.mobygames.com/game/disney-planes", 
"http://www.mobygames.com/game/gold-rush-the-game", 
"http://www.mobygames.com/game/pirate-plague-of-the-dead", 
"http://www.mobygames.com/game/disney-planes-fire-rescue", 
"http://www.mobygames.com/game/city-car-driving", 
"http://www.mobygames.com/game/homebrew", 
"http://www.mobygames.com/game/my-summer-car", 
"http://www.mobygames.com/game/air-race", 
"http://www.mobygames.com/game/hot-rod-racer", 
"http://www.mobygames.com/game/cockpit", 
"http://www.mobygames.com/game/eisenbahn-x", 
"http://www.mobygames.com/game/professional-farmer-2014", 
"http://www.mobygames.com/game/professional-lumberjack-2015", 
"http://www.mobygames.com/game/mining-and-tunneling-simulator", 
"http://www.mobygames.com/game/snowcat-simulator-2014", 
"http://www.mobygames.com/game/plant-fire-department-the-simulation", 
"http://www.mobygames.com/game/safety-driving-car", 
"http://www.mobygames.com/game/safety-driving-motorbike", 
"http://www.mobygames.com/game/arctic-trucker-simulator", 
"http://www.mobygames.com/game/warehouse-and-logistics-simulator", 
"http://www.mobygames.com/game/driving-school-simulator", 
"http://www.mobygames.com/game/war-truck-simulator", 
"http://www.mobygames.com/game/sailaway-the-sailing-simulator", 
"http://www.mobygames.com/game/monster-bus-snocat", 
"http://www.mobygames.com/game/wrecking-ball-machine", 
"http://www.mobygames.com/game/epic-roller-coasters", 
"http://www.mobygames.com/game/pure-farming-2018", 
"http://www.mobygames.com/game/subway-simulator", 
"http://www.mobygames.com/game/realflight-8", 
"http://www.mobygames.com/game/eisenbahnwelten-interaktiv", 
"http://www.mobygames.com/game/farm-expert-2018-mobile", 
"http://www.mobygames.com/game/high-speed-trains_", 
"http://www.mobygames.com/game/vr-stock-car-racers", 
"http://www.mobygames.com/game/rfactor-2", 
"http://www.mobygames.com/game/rogue-system", 
"http://www.mobygames.com/game/frmula-truck-simulador-oficial-2013", 
"http://www.mobygames.com/game/copa-petrobras-de-marcas", 
"http://www.mobygames.com/game/esail", 
"http://www.mobygames.com/game/lunar-lander____", 
"http://www.mobygames.com/game/disaster-response-unit-thw", 
"http://www.mobygames.com/game/armor-contest", 
"http://www.mobygames.com/game/v-racer-hoverbike", 
"http://www.mobygames.com/game/darkest-hour-europe-44-45", 
"http://www.mobygames.com/game/city-bus-simulator-2018", 
"http://www.mobygames.com/game/mulle-mecks-cars", 
"http://www.mobygames.com/game/mulle-mecks-boats", 
"http://www.mobygames.com/game/rolling-line", 
"http://www.mobygames.com/game/lem___", 
"http://www.mobygames.com/game/ghost-mountain-roller-coaster", 
"http://www.mobygames.com/game/helicopter-flight-simulator", 
"http://www.mobygames.com/game/panekit", 
"http://www.mobygames.com/game/acro-fs", 
"http://www.mobygames.com/game/bygg-btar-med-mulle-meck-specialversion", 
"http://www.mobygames.com/game/motogp-18", 
"http://www.mobygames.com/game/flashing-lights-police-fire-ems", 
"http://www.mobygames.com/game/bus-simulator-18", 
"http://www.mobygames.com/game/dirt-bike-insanity", 
"http://www.mobygames.com/game/police-helicopter", 
"http://www.mobygames.com/game/flug-simulator-fr-airbus-a-320-europa-edition-version-2", 
"http://www.mobygames.com/game/trigger__", 
"http://www.mobygames.com/game/thomas-friends-magic-tracks", 
"http://www.mobygames.com/game/thomas-friends-minis", 
"http://www.mobygames.com/game/thomas-friends-race-on", 
"http://www.mobygames.com/game/thomas-friends-express-delivery", 
"http://www.mobygames.com/game/speedway-99", 
"http://www.mobygames.com/game/lunar-lander_____", 
"http://www.mobygames.com/game/asphalt-3-3d-street-rules", 
"http://www.mobygames.com/game/lunar-lander______", 
"http://www.mobygames.com/game/asphalt-4-elite-racing_", 
"http://www.mobygames.com/game/lunar-lander_______", 
"http://www.mobygames.com/game/lunar-lander________", 
"http://www.mobygames.com/game/asphalt-6-adrenaline_", 
"http://www.mobygames.com/game/asphalt-6-adrenaline__", 
"http://www.mobygames.com/game/lunar-lander_________", 
"http://www.mobygames.com/game/rocket__", 
"http://www.mobygames.com/game/lunar-landing-game", 
"http://www.mobygames.com/game/sailplane", 
"http://www.mobygames.com/game/moonlanding", 
"http://www.mobygames.com/game/f1-2018", 
"http://www.mobygames.com/game/schrottplatz-simulator-2011", 
"http://www.mobygames.com/game/ferrari-gt-evolution_", 
"http://www.mobygames.com/game/nintendo-labo-toy-con-03-vehicle-kit-", 
"http://www.mobygames.com/game/moon-landing", 
"http://www.mobygames.com/game/mars-landing", 
"http://www.mobygames.com/game/woodcutter-simulator-2011-multiplayer-edition", 
"http://www.mobygames.com/game/xlander", 
"http://www.mobygames.com/game/lunar-module", 
"http://www.mobygames.com/game/kim-goes-to-the-moon", 
"http://www.mobygames.com/game/basic-software-library-volume-i-business-and-recreational", 
"http://www.mobygames.com/game/construction-simulator_", 
"http://www.mobygames.com/game/startrek", 
"http://www.mobygames.com/game/v-rally-4", 
"http://www.mobygames.com/game/bus-driver-simulator-19", 
"http://www.mobygames.com/game/nascar-heat-3", 
"http://www.mobygames.com/game/atv-quadracer-ultimate", 
"http://www.mobygames.com/game/space-flight", 
"http://www.mobygames.com/game/vox-machinae", 
"http://www.mobygames.com/game/choro-q-wonderful", 
"http://www.mobygames.com/game/zamboni-challenge", 
"http://www.mobygames.com/game/frontier-pilot-simulator", 
"http://www.mobygames.com/game/farming-simulator-19", 
"http://www.mobygames.com/game/cityconomy-service-for-your-city", 
"http://www.mobygames.com/game/dires-gigerloop", 
"http://www.mobygames.com/game/vip-lunar-lander", 
"http://www.mobygames.com/game/conduct-this", 
"http://www.mobygames.com/game/conduct-ar", 
"http://www.mobygames.com/game/mars-lander", 
"http://www.mobygames.com/game/overtake", 
"http://www.mobygames.com/game/farming-simulator-19-c64-edition", 
"http://www.mobygames.com/game/tourist-bus-simulator", 
"http://www.mobygames.com/game/flywings-2018", 
"http://www.mobygames.com/game/flying-aces-navy-pilot-simulator", 
"http://www.mobygames.com/game/vr-flight-simulator-new-york-cessna", 
"http://www.mobygames.com/game/red-bull-air-race-2", 
"http://www.mobygames.com/game/tonka-on-the-job", 
"http://www.mobygames.com/game/orbitral", 
"http://www.mobygames.com/game/climbin-man", 
"http://www.mobygames.com/game/real-car-driving-experience", 
"http://www.mobygames.com/game/fa-18-hornet_", 
"http://www.mobygames.com/game/x-plane-10", 
"http://www.mobygames.com/game/space-shuttle_", 
"http://www.mobygames.com/game/bloodhound-ssc", 
"http://www.mobygames.com/game/stradale-racing-simulator", 
"http://www.mobygames.com/game/space-simulator", 
"http://www.mobygames.com/game/picasim", 
"http://www.mobygames.com/game/crash-landing", 
"http://www.mobygames.com/game/pizza-parking", 
"http://www.mobygames.com/game/lunarlanderti", 
"http://www.mobygames.com/game/stunt-flyer", 
"http://www.mobygames.com/game/dirt-rally-20", 
"http://www.mobygames.com/game/sega-rally-online-arcade", 
"http://www.mobygames.com/game/gpak-2", 
"http://www.mobygames.com/game/sands-of-mars", 
"http://www.mobygames.com/game/cubix-robots-for-everyone-race-n-robots_", 
"http://www.mobygames.com/game/dragon-tales-dragon-wings", 
"http://www.mobygames.com/game/sfs-walletsize-spaceship-simulator", 
"http://www.mobygames.com/game/mission-space", 
"http://www.mobygames.com/game/jet-copter-x", 
"http://www.mobygames.com/game/motogp-19", 
"http://www.mobygames.com/game/level-ii-basic-reference-manual-included-games", 
"http://www.mobygames.com/game/f1-2019-anniversary-edition", 
"http://www.mobygames.com/game/poly-88-included-games", 
"http://www.mobygames.com/game/driving-school-2017", 
"http://www.mobygames.com/game/driving-school-2016", 
"http://www.mobygames.com/game/poly-runner-vr", 
"http://www.mobygames.com/game/air-brigade", 
"http://www.mobygames.com/game/world-of-riders", 
"http://www.mobygames.com/game/fireground", 
"http://www.mobygames.com/game/radio-helicopter-ii", 
"http://www.mobygames.com/game/forklift-simulator", 
"http://www.mobygames.com/game/logistics-expert", 
"http://www.mobygames.com/game/parking-3d", 
"http://www.mobygames.com/game/forklift-simulator-2019", 
"http://www.mobygames.com/game/forklift-the-simulation", 
"http://www.mobygames.com/game/symphony-of-stars", 
"http://www.mobygames.com/game/battle-killer-stuka", 
"http://www.mobygames.com/game/f1-mobile-racing", 
"http://www.mobygames.com/game/taxi__", 
"http://www.mobygames.com/game/european-ship-simulator", 
"http://www.mobygames.com/game/grid___", 
"http://www.mobygames.com/game/need-for-speed-underground-j-tune", 
"http://www.mobygames.com/game/rf9-realflight", 
"http://www.mobygames.com/game/midnight-landing", 
"http://www.mobygames.com/game/top-landing", 
"http://www.mobygames.com/game/nascar-heat-4", 
"http://www.mobygames.com/game/demolish-build-17", 
"http://www.mobygames.com/game/lunar-lander-for-the-6502", 
"http://www.mobygames.com/game/ultra-off-road-2019-alaska", 
"http://www.mobygames.com/game/lunarlanderapple", 
"http://www.mobygames.com/game/flightdeck", 
"http://www.mobygames.com/game/mountain-rescue-simulator", 
"http://www.mobygames.com/game/truck-simulator-usa", 
"http://www.mobygames.com/game/need-for-speed-undercover_____", 
"http://www.mobygames.com/game/car-parking-simulator", 
"http://www.mobygames.com/game/rds-russian-drift-series-the-official-drift-videogame", 
"http://www.mobygames.com/game/emergency-call-ambulance", 
"http://www.mobygames.com/game/farming-simulator-20", 
"http://www.mobygames.com/game/super-speed", 
"http://www.mobygames.com/game/derail-valley", 
"http://www.mobygames.com/game/ironwolf-vr", 
"http://www.mobygames.com/game/juggernaut_", 
"http://www.mobygames.com/game/cassette-28-cockpit", 
"http://www.mobygames.com/game/japanese-rail-sim-journey-to-kyoto", 
"http://www.mobygames.com/game/tetsud-nippon-rosen-tabiki-kan-sha-thomas-hen-igawa-tetsud-o-has", 
"http://www.mobygames.com/game/super-speed-christmas-edition", 
"http://www.mobygames.com/game/leisure-library", 
"http://www.mobygames.com/game/gunship-battle2-vr", 
"http://www.mobygames.com/game/night-flight__", 
"http://www.mobygames.com/game/ridge-racer-2_", 
"http://www.mobygames.com/game/rave-racer", 
"http://www.mobygames.com/game/extreme-trucks-simulator", 
"http://www.mobygames.com/game/ground-control-simulator", 
"http://www.mobygames.com/game/bike-daisuki-hashiriya-tamash-riders-spirits", 
"http://www.mobygames.com/game/duel-test-drive-ii_", 
"http://www.mobygames.com/game/duel-test-drive-ii__", 
"http://www.mobygames.com/game/space-shuttle__", 
"http://www.mobygames.com/game/hello-mars", 
"http://www.mobygames.com/game/mars-odyssey", 
"http://www.mobygames.com/game/rc-plane-vr", 
"http://www.mobygames.com/game/trainervr", 
"http://www.mobygames.com/game/rover-mechanic-simulator", 
"http://www.mobygames.com/game/need-for-speed-hot-pursuit-2_", 
"http://www.mobygames.com/game/360-hover-parking", 
"http://www.mobygames.com/game/police-cars-parking", 
"http://www.mobygames.com/game/shopping-mall-parking", 
"http://www.mobygames.com/game/overflight", 
"http://www.mobygames.com/game/tony-stewarts-sprint-car-racing", 
"http://www.mobygames.com/game/coaster_", 
"http://www.mobygames.com/game/dcl-the-game", 
"http://www.mobygames.com/game/farmer-sim-2018", 
"http://www.mobygames.com/game/flight-sim-18", 
"http://www.mobygames.com/game/over-drive", 
"http://www.mobygames.com/game/apocalypse-rider", 
"http://www.mobygames.com/game/top-speed-drag-fast-racing", 
"http://www.mobygames.com/game/top-speed-2-racing-legends", 
"http://www.mobygames.com/game/giant-trucks-driving-simulator", 
"http://www.mobygames.com/game/ford-racing-full-blown", 
"http://www.mobygames.com/game/chase-hq-plus-sci", 
"http://www.mobygames.com/game/motogp-20", 
"http://www.mobygames.com/game/omg-police-car-chase-tv-simulator", 
"http://www.mobygames.com/game/kamikaze-cabbie", 
"http://www.mobygames.com/game/snowrunner", 
"http://www.mobygames.com/game/carnage-battle-arena", 
"http://www.mobygames.com/game/car-driving-school-simulator", 
"http://www.mobygames.com/game/forking-hell", 
"http://www.mobygames.com/game/bus-mechanic-simulator", 
"http://www.mobygames.com/game/vr-motor-racing-mania", 
"http://www.mobygames.com/game/wildtrax-racing", 
"http://www.mobygames.com/game/taxi-sim", 
"http://www.mobygames.com/game/drag-sim-2018", 
"http://www.mobygames.com/game/jumbo", 
"http://www.mobygames.com/game/bob-bishops-hires-rocket-lander", 
"http://www.mobygames.com/game/ship-sim-2019", 
"http://www.mobygames.com/game/four-trax", 
"http://www.mobygames.com/game/power-surge", 
"http://www.mobygames.com/game/rc-airsim", 
"http://www.mobygames.com/game/air-flight-simulation", 
"http://www.mobygames.com/game/truck-logistics-simulator", 
"http://www.mobygames.com/game/space-ace-21", 
"http://www.mobygames.com/game/car-caramba-driving-simulator", 
"http://www.mobygames.com/game/gas-station-2-highway-service", 
"http://www.mobygames.com/game/landing-gear", 
"http://www.mobygames.com/game/nascar-heat-5", 
"http://www.mobygames.com/game/harley-davidson-la-riders", 
"http://www.mobygames.com/game/city-bus-driving-sim", 
"http://www.mobygames.com/game/microsoft-flight-simulator_", 
"http://www.mobygames.com/game/crosscountry-bc", 
"http://www.mobygames.com/game/icarus-starship-command-simulator", 
"http://www.mobygames.com/game/basic-and-intermediate-lunar-lander", 
"http://www.mobygames.com/game/drone-racing-league-simulator", 
"http://www.mobygames.com/game/space-trek-ii", 
"http://www.mobygames.com/game/trucker-parking-simulator-2-a-real-monster-truck-lorry-driving-t", 
"http://www.mobygames.com/game/bh-trials-digger-parkour", 
"http://www.mobygames.com/game/tho-simulator", 
"http://www.mobygames.com/game/blonde-driver", 
"http://www.mobygames.com/game/bus-driver-simulator", 
"http://www.mobygames.com/game/tracks-toybox-edition", 
"http://www.mobygames.com/game/extreme-forklifting-2", 
"http://www.mobygames.com/game/into-space", 
"http://www.mobygames.com/game/flughafen-feuerwehr-simulator-2013", 
"http://www.mobygames.com/game/warbirds-dawn-of-aces", 
"http://www.mobygames.com/game/street-kart-racing", 
"http://www.mobygames.com/game/great-air-battles-1941-1942", 
"http://www.mobygames.com/game/tc2000-racing-20", 
"http://www.mobygames.com/game/richard-burns-rally_", 
"http://www.mobygames.com/game/beach-buggy-racing-2", 
"http://www.mobygames.com/game/cargo-crew-port-truck-driver", 
"http://www.mobygames.com/game/sci-fi-sampler", 
"http://www.mobygames.com/game/moorhuhn-kart-2", 
"http://www.mobygames.com/game/motogp-21", 
"http://www.mobygames.com/game/fx-racer", 
"http://www.mobygames.com/game/trs-80-lunar-lander", 
"http://www.mobygames.com/game/space-trek-i", 
"http://www.mobygames.com/game/caterpillar-construction-tycoon", 
"http://www.mobygames.com/game/mountain-pilot", 
"http://www.mobygames.com/game/nascar-99-legacy", 
"http://www.mobygames.com/game/vertigo___", 
"http://www.mobygames.com/game/space-voyage_", 
"http://www.mobygames.com/game/fly-this-flight-control-tower", 
"http://www.mobygames.com/game/zarbor", 
"http://www.mobygames.com/game/highway-haste", 
"http://www.mobygames.com/game/taxi-driver-the-simulation", 
"http://www.mobygames.com/game/trucksimulation-16", 
"http://www.mobygames.com/game/startrektaylor", 
"http://www.mobygames.com/game/spaceship-cannon", 
"http://www.mobygames.com/game/moon-lander", 
"http://www.mobygames.com/game/spacetrek", 
"http://www.mobygames.com/game/trek-x", 
"http://www.mobygames.com/game/worlds-best-star-trek", 
"http://www.mobygames.com/game/real-time-lunar-lander", 
"http://www.mobygames.com/game/startrek_", 
"http://www.mobygames.com/game/multi-floor-garage-driver", 
"http://www.mobygames.com/game/startrekpetshack", 
"http://www.mobygames.com/game/pet-trek-3", 
"http://www.mobygames.com/game/train-simulator-london-subway", 
"http://www.mobygames.com/game/garage-master", 
"http://www.mobygames.com/game/moto-rider-go-highway-traffic", 
"http://www.mobygames.com/game/star-trek-2001", 
"http://www.mobygames.com/game/lunarlanderab", 
"http://www.mobygames.com/game/moonlander_", 
"http://www.mobygames.com/game/star-trek-iv", 
"http://www.mobygames.com/game/metro-simulator", 
"http://www.mobygames.com/game/shopping-mall-parking-lot", 
"http://www.mobygames.com/game/f1-world-championship-edition_", 
"http://www.mobygames.com/game/jrgs-flight-sim", 
"http://www.mobygames.com/game/spacewarwlgregoryjr", 
"http://www.mobygames.com/game/star-trek-advanced", 
"http://www.mobygames.com/game/racing-xtreme-2-top-monster-truck-offroad-fun", 
"http://www.mobygames.com/game/truck-driver-depot-parking", 
"http://www.mobygames.com/game/driving-school-sim-2020", 
"http://www.mobygames.com/game/car-physics-simulator", 
"http://www.mobygames.com/game/race-driving-license-test", 
"http://www.mobygames.com/game/lawn-mowing-simulator", 
"http://www.mobygames.com/game/ultimate-bumper-cars-dodgems", 
"http://www.mobygames.com/game/roundabout-2-city-driving-sim", 
"http://www.mobygames.com/game/speedway-championships", 
"http://www.mobygames.com/game/sttr2_", 
"http://www.mobygames.com/game/startrekunivac4", 
"http://www.mobygames.com/game/bus-simulator-21", 
"http://www.mobygames.com/game/lake_", 
"http://www.mobygames.com/game/univac-startrek", 
"http://www.mobygames.com/game/moon-base_", 
"http://www.mobygames.com/game/nascar-challenge", 
"http://www.mobygames.com/game/nascar-heat-ultimate-edition", 
"http://www.mobygames.com/game/farming-simulator-22", 
"http://www.mobygames.com/game/crash-city-heavy-traffic-drive", 
"http://www.mobygames.com/game/coastline-flight-simulator"]

coreGameRelease = []


for site in moby_urls:
    response = requests.get(site)
    print(response)
    soup = BeautifulSoup(response.content, 'lxml')
    smallsoup = soup.find('div', {'id': 'coreGameRelease'}).text
    coreGameRelease.append(smallsoup)
    sleep(10.01)

len(coreGameRelease)

### when you return, run cell below

import pickle
f = open("../data/release_info_from_webscrape_not_api.pkl","wb")
pickle.dump(coreGameRelease,f)
f.close()

clean_coreGameRelease = []

#clean list
for line in coreGameRelease:
    line = line.replace('\xa0', ' ')
    clean_coreGameRelease.append(line)

clean_coreGameRelease
