==> notebooks/7 - dropping racing.ipynb <==
import pandas as pd
import numpy as np

#example from https://stackoverflow.com/questions/47914985/drop-a-row-in-a-pandas-dataframe-if-any-column-contains-a-certain-value-in-said

# example df

#     a   b   c   d   e   f   g   h   i   j
# 0  95  79  14  64  97   3   0  49  84   2
# 1  52  82  72  40  55  26  21  69   2  35
# 2  30  63  27  63  14  23  70  62  83  80
# 3  23  90  96  43  75  23  27  47  83  40
# 4  82  87  63  37  36  58  88  18  50  29
# 5  92  29  93   6  62  85  87  74  21  32
# 6  23  64  19  22  50  86  51  93  46   1
# 7  56   0   8  46   1  56  82  15  55  45
# 8  67  35  83  65  63  58  48  85  75  90
# 9  12   7  45  28  33  98  42  93  78  12

# let's drop all rows containing at least one 0:

# In [10]: df.eq(0)
# Out[10]:
#        a      b      c      d      e      f      g      h      i      j
# 0  False  False  False  False  False  False   True  False  False  False
# 1  False  False  False  False  False  False  False  False  False  False
# 2  False  False  False  False  False  False  False  False  False  False
# 3  False  False  False  False  False  False  False  False  False  False
# 4  False  False  False  False  False  False  False  False  False  False
# 5  False  False  False  False  False  False  False  False  False  False
# 6  False  False  False  False  False  False  False  False  False  False
# 7  False   True  False  False  False  False  False  False  False  False
# 8  False  False  False  False  False  False  False  False  False  False
# 9  False  False  False  False  False  False  False  False  False  False

# In [11]: res = df[~df.eq(0).any(1)]

# In [12]: res
# Out[12]:
#     a   b   c   d   e   f   g   h   i   j
# 1  52  82  72  40  55  26  21  69   2  35
# 2  30  63  27  63  14  23  70  62  83  80
# 3  23  90  96  43  75  23  27  47  83  40
# 4  82  87  63  37  36  58  88  18  50  29
# 5  92  29  93   6  62  85  87  74  21  32
# 6  23  64  19  22  50  86  51  93  46   1
# 8  67  35  83  65  63  58  48  85  75  90
# 9  12   7  45  28  33  98  42  93  78  12

df = pd.read_csv("../data/all_genres_before_racing_drop_17_Dec_21.csv")

df.eq("Racing / driving")

no_racing = df[~df.eq("Racing / driving").any(1)]

no_racing.to_csv("../data/no_racing.csv")

only_racing = df[df.eq("Racing / driving").any(1)]

only_racing.to_csv("../data/only_racing.csv")

### doing some groupbys for genre

res = df.groupby(["genre_name6"])["title"].nunique().sort_values(ascending=False)
res.to_csv("../data/groupby_genre6.csv")

### looping through columns to get list of genre tags

genre_string_list =[]

for data in df["genre_name7"]:
    if data not in genre_string_list:
        genre_string_list.append(data)


# for (columnName, columnData) in df.iteritems():
#     print(type(columnData.values))
#     for string in columnData.values():
#         if string not in genre_string_list:
#             genre_string_list.append(string)

len(genre_string_list)

dataframe_lens = []

### looping through with eq method

for string in genre_string_list:
    resdf = df[df.eq(string).any(1)]
    dataframe_lens.append(len(resdf))
    

len(dataframe_lens)

genres_df = pd.DataFrame()

genres_df["genre_name"] = genre_string_list

genres_df["count_of_genre"] = dataframe_lens

genres_df.to_csv("../data/genre_tag_counts.csv")
